server/web/requirements.txt
# Python 3.10.14
flask==3.0.3
loguru==0.7.2
pytz==2024.1
flask_sqlalchemy==3.1.1
pyperclip==1.9.0